src/modules/students/services/DeleteStudentService.ts
import { injectable, inject } from 'tsyringe';

import IStudentRepository from '../repositories/IStudentRepository';
import HttpRequestError from '../../../shared/errors/HttpRequestError';

@injectable()
class DeleteStudentService {
  private studentRepository: IStudentRepository;

  constructor(
    @inject('StudentRepository')
    studentRepository: IStudentRepository,
  ) {
    this.studentRepository = studentRepository;
  }

  public async execute(studentId: string): Promise<void> {
    const student = await this.studentRepository.findById(studentId);
    if (!student) {
      throw new HttpRequestError('Invalid student id');
    }

    await this.studentRepository.delete(studentId);
  }
}

export default DeleteStudentService;
